+++ packages/components/src/components/@shared/_kol-select-mixin.scss
@@ -1,0 +1,38 @@
+@use '../../styles/global' as *;
+@use '../../styles/kol-alert-mixin' as *;
+@use '../../styles/kol-form-field-mixin' as *;
+@use '../../styles/kol-input-container-mixin' as *;
+@use '../../styles/kol-input-mixin' as *;
+@use '../@shared/mixins' as *;
+
+@mixin kol-select {
+	@include kol-alert;
+	@include kol-form-field;
+	@include kol-input-container;
+
+	.kol-select {
+		background-color: transparent;
+		width: 100%;
+
+		&:not(:disabled) {
+			cursor: pointer;
+		}
+
+		&:not([multiple], [size]) {
+			height: 2.75em;
+		}
+
+		&:focus {
+			outline: none;
+		}
+
+		&__option:checked::before {
+			content: '✓ ';
+		}
+
+		/* needed hack for vertical alignment */
+		&[multiple] option {
+			padding: calc((var(--a11y-min-size) - to-rem(16)) / 2) 0.5em;
+		}
+	}
+}
