+++ packages/components/src/components/pagination/style.scss
@@ -2,7 +2,9 @@
 @use '../../styles/global' as *;
 @use '../tooltip/style.scss' as *;
 @use '../@shared/kol-pagination-mixin' as *;
+@use '../@shared/kol-select-mixin' as *;
 
 @layer kol-component {
 	@include kol-pagination-styles;
+	@include kol-select;
 }
